requirements.txt
cachelib==0.10.2
imageio==2.31.1
librosa==0.8.0
#numpy==1.21.6
# tf依赖numpy 1.19.5, 安装此文件，可以不需要指定numpy，会默认被安装依赖1.19.5的版本
#numpy==1.19.5
# opencv_python==4.5.2.54
# Pillow==10.0.0
psutil==5.8.0
python_speech_features==0.6
resampy==0.4.2
scikit_image==0.19.3
scipy==1.7.0
tensorflow-gpu==2.6.0
# torch==1.13.1
# torchvision==0.14.1
tqdm==4.62.2
ttskit==0.3.8
pyttsx4==3.0.15
websockets==11.0.3
opencv-python==4.8.0.74
pika